Delete unneccessary vet mapping, merge branch with main
