+++ api/src/main/java/SpringProject/_Spring/controller/AdminpageController.java
@@ -3,8 +3,8 @@ package SpringProject._Spring.controller;
 import SpringProject._Spring.dto.ApiResponse;
 import SpringProject._Spring.dto.authentication.client.ClientMapping;
 import SpringProject._Spring.dto.authentication.client.ClientPageResponseDTO;
+import SpringProject._Spring.dto.authentication.vet.VetMapping;
 import SpringProject._Spring.dto.authentication.vet.VetPageResponseDTO;
-import SpringProject._Spring.dto.vet.VetMapping;
 import SpringProject._Spring.model.authentication.Client;
 import SpringProject._Spring.model.authentication.Vet;
 import SpringProject._Spring.service.authentication.ClientService;
